The implementation of a shape has been modified in such a way, that the array of types and part starts do not follow each other anymore. Instead a new struct (atlas_shape_part_info_s) has been introduced, that holds the type of a shape and the starting index for the coordinates.
In parallel the getter functions for the type of a part and the pan part start had to be modified.

Also the copying of the coordinates when creating a shape had to be modified, since the count for the pointer is calculated differently. That's also due to the introduction of the new struct.

static_data has been renamed to shape_data, since the static characteristic is lost.

Furthermore some empty methods for what is still to do have been written down.

Finally the function comparing two shapes for equality has been written together with a test case.