Test case files have been moved into test directory.
atlas_shape_line_intersect has been modified, such that if there is an intersection the coordinate is written and 1 returned. If 1 is returned and no coordinate both lines are on top on each other and 0 if there is no intersection.

Also atlas_shape_line_intersect was moved to a seperate geometry file.